# A constructor for BoundedArray