Fix #1494.