@pleroy 2017-08-01
Confirmed that the benchmarks are unaffected by this change:

Before:
```
Benchmark                                            Time           CPU Iterations
----------------------------------------------------------------------------------
BM_VisibleSegmentsOrbit/10                        1005 ns       1001 ns     747858 average visible segments: 1.000000
BM_VisibleSegmentsOrbit/100                       6288 ns       6258 ns     112179 average visible segments: 0.980000
BM_VisibleSegmentsOrbit/1000                     60347 ns      59280 ns      10000 average visible segments: 0.966000
BM_VisibleSegmentsRandomEverywhere/1000          61191 ns      58407 ns      11218 average visible segments: 1.073000
BM_VisibleSegmentsRandomNoIntersection/1000      44133 ns      44293 ns      17258 average visible segments: 1.000000
```

After:
```
Benchmark                                            Time           CPU Iterations
----------------------------------------------------------------------------------
BM_VisibleSegmentsOrbit/10                         975 ns        980 ns     747858 average visible segments: 1.000000
BM_VisibleSegmentsOrbit/100                       6143 ns       6119 ns     112179 average visible segments: 0.980000
BM_VisibleSegmentsOrbit/1000                     59606 ns      59797 ns      11218 average visible segments: 0.966000
BM_VisibleSegmentsRandomEverywhere/1000          60751 ns      58407 ns      11218 average visible segments: 1.073000
BM_VisibleSegmentsRandomNoIntersection/1000      44086 ns      43804 ns      16026 average visible segments: 1.000000
```